Split long sections into subdirectories.
I've found, that wikibook has just part of Section 5, so I converted
missing text from html and put it on the wikibook. I put here the
conversion script just in case. Probably we can reuse it if we
will work on restoring links present in HTML, but lost on wikibook.